tools/sdk-ai-bots/Embeddings/requirements.txt
azure-identity==1.16.1
azure-search-documents==11.6.0b2
langchain==0.1.8
openai==0.27.8
tiktoken==0.4.0
python-dotenv==0.19.1